Add TaskCompletionSourceWithCancellation implementation. Move WithCancellation extension method to dotnet-monitor.
