+++ src/Tests/Microsoft.Diagnostics.Monitoring.TestCommon/Runners/DotNetRunner.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information
 
+using Microsoft.Diagnostics.Tools.Monitor;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -174,12 +175,7 @@ namespace Microsoft.Diagnostics.Monitoring.TestCommon.Runners
         {
             if (HasStarted && !_process.HasExited)
             {
-                TaskCompletionSource<object> cancellationSource = new(TaskCreationOptions.RunContinuationsAsynchronously);
-                using IDisposable _ = token.Register(() => cancellationSource.TrySetCanceled(token));
-
-                await Task.WhenAny(
-                    ExitedTask,
-                    cancellationSource.Task).Unwrap();
+                await ExitedTask.WithCancellation(token);
             }
         }
