+++ src/Tools/dotnet-monitor/TaskExtensions.cs
@@ -2,46 +2,17 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
 using System.Threading;
 using System.Threading.Tasks;
-using Xunit.Abstractions;
 
+#if UNITTEST
 namespace Microsoft.Diagnostics.Monitoring.TestCommon
+#else
+namespace Microsoft.Diagnostics.Tools.Monitor
+#endif
 {
     public static class TaskExtensions
     {
-        public static async Task SafeAwait(this Task task, ITestOutputHelper outputHelper = null)
-        {
-            if (task != null)
-            {
-                try
-                {
-                    await task.ConfigureAwait(false);
-                }
-                catch (Exception ex)
-                {
-                    outputHelper?.WriteLine("Warning: Exception thrown while awaiting task: {0}", ex);
-                }
-            }
-        }
-
-        public static async Task<T> SafeAwait<T>(this Task<T> task, ITestOutputHelper outputHelper = null, T fallbackValue = default(T))
-        {
-            if (task != null)
-            {
-                try
-                {
-                    return await task.ConfigureAwait(false);
-                }
-                catch (Exception ex)
-                {
-                    outputHelper?.WriteLine("Warning: Exception thrown while awaiting task: {0}", ex);
-                }
-            }
-            return fallbackValue;
-        }
-
         public static async Task WithCancellation(this Task task, CancellationToken token)
         {
             using CancellationTokenSource localTokenSource = CancellationTokenSource.CreateLinkedTokenSource(token);
